% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{add_fourier}
\alias{add_fourier}
\title{Add fourier terms to tsibble}
\usage{
add_fourier(original, K_fourier = Inf)
}
\arguments{
\item{original}{original tsibble}

\item{K_fourier}{number of fourier terms}
}
\value{
tsibble with fourier terms
}
\description{
Add fourier terms to tsibble
}
\details{
Add fourier terms to tsibble
}
\examples{
test_ts = stats::ts(rnorm(100), start = c(2000, 1), freq = 12)
test_tsibble = tsibble::as_tsibble(test_ts)
add_fourier(test_tsibble)
}
