% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_processing.R
\name{add_hf_lags}
\alias{add_hf_lags}
\title{Add high frequency lags to low frequency tsibble}
\usage{
add_hf_lags(lf_data, hf_data, hf_variable = NULL, n_lags = 1,
  one_row = TRUE)
}
\arguments{
\item{lf_data}{low frequency data}

\item{hf_data}{high frequency data}

\item{hf_variable}{names of high frequency time series, all by default}

\item{n_lags}{number of lags to add}

\item{one_row}{output representation}
}
\value{
tsibble with high frequency lags added
}
\description{
Add high frequency lags to low frequency tsibble
}
\details{
Add high frequency lags to low frequency tsibble.
}
\examples{
# no yet
}
