% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{add_lags}
\alias{add_lags}
\title{Add lags of many variables}
\usage{
add_lags(original, variable_names, lags = c(1, 2))
}
\arguments{
\item{original}{original tsibble}

\item{variable_names}{variables to add lags, with quotes! Like "gdp" and not gdp.}

\item{lags}{desired lags, a vector}
}
\value{
tsibble with lags of specified variables
}
\description{
Add lags of many variables
}
\details{
Add lags of many variables
The name of variables should be with quotes.
Designed mainly for explanatory variables.
}
\examples{
test_ts = stats::ts(rnorm(100), start = c(2000, 1), freq = 12)
test_tsibble = tsibble::as_tsibble(test_ts)
add_lags(test_tsibble, "value", 1:7)
}
