% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{add_point_forecasts}
\alias{add_point_forecasts}
\title{Add point forecast to models tibble}
\usage{
add_point_forecasts(model_list_fitted)
}
\arguments{
\item{model_list_fitted}{tibble with one model per row}
}
\value{
original tibble plus point forecasts
}
\description{
Add point forecast to models tibble
}
\details{
Add point forecast to models tibble.
}
\examples{
# no yet
}
