% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{add_trend}
\alias{add_trend}
\title{Add linear and root trends to tibble}
\usage{
add_trend(original)
}
\arguments{
\item{original}{tibble}
}
\value{
tibble with trend_lin and trend_root columns
}
\description{
Add linear and root trends to tibble
}
\details{
Add linear and root trends to tibble
}
\examples{
# dumb example: add trend to cross section :) :)
add_trend(cars)
}
