% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Almon_lag.R
\name{Almon_lag}
\alias{Almon_lag}
\title{Constructs Almon lag polynomial}
\usage{
Almon_lag(polydegree, C, R)
}
\arguments{
\item{polydegree}{degree of polynomial}

\item{C}{number of columns in Q matrix (???)}

\item{R}{two numbers (???)}
}
\value{
Q matrix
}
\description{
Constructs Almon lag polynomial
}
\details{
Function constructs Almon lag polynomial as in the article by
Matteo Mogliani Bayesian MIDAS Penalized Regressions:
Estimation, Selection, and Prediction 2 Jan 2020
Enables to put endpoint restrictions on the value and on slope of the lag polynomial default R=c(1,1)
Meaning that restrictions jointly constrain the weighting structure to tail off slowly to zero
}
\examples{
Almon_lag(2, 5, c(1, 0))
}
