% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{arima101_010_fun}
\alias{arima101_010_fun}
\title{Do forecast using ARIMA(1,0,1)-SARIMA(0,1,0)}
\usage{
arima101_010_fun(model_sample, h, target = "value")
}
\arguments{
\item{model_sample}{preferably tsibble}

\item{h}{forecasting horizon, is ignored}

\item{target}{name of the target variable, "value" by default}
}
\value{
ARIMA(1,0,1)-SARIMA(0,1,0) model
}
\description{
Do forecast using ARIMA(1,0,1)-SARIMA(0,1,0)
}
\details{
Do forecast using ARIMA(1,0,1)-SARIMA(0,1,0)
}
\examples{
test = dplyr::tibble(date = as.Date("2017-01-01") + 0:9, value = rnorm(10))
arima101_010_fun(test, 1)
}
