% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{arima101_101_fun}
\alias{arima101_101_fun}
\title{Do forecast using ARIMA(1,0,1)-SARIMA(1,0,1)}
\usage{
arima101_101_fun(model_sample, h = 0, target = "value")
}
\arguments{
\item{model_sample}{preferably tsibble with "value" column}

\item{h}{forecasting horizon, is ignored}

\item{target}{name of the target variable, "value" by default}
}
\value{
ARIMA(1,0,1)-SARIMA(1,0,1) model
}
\description{
Do forecast using ARIMA(1,0,1)-SARIMA(1,0,1)
}
\details{
Do forecast using ARIMA(1,0,1)-SARIMA(1,0,1)
}
\examples{
test = dplyr::tibble(date = as.Date("2017-01-01") + 0:9, value = rnorm(10))
arima101_101_fun(test, 1)
}
