% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{arima11_fun}
\alias{arima11_fun}
\title{Do forecast using ARIMA(1,0,1)-SARIMA[12](1,0,1)}
\usage{
arima11_fun(model_sample, h)
}
\arguments{
\item{model_sample}{preferably tsibble with "value" column}

\item{h}{forecasting horizon, is ignored}
}
\value{
ARIMA(1,0,1)-SARIMA[12](1,0,1) model
}
\description{
Do forecast using ARIMA(1,0,1)-SARIMA[12](1,0,1)
}
\details{
Do forecast using ARIMA(1,0,1)-SARIMA[12](1,0,1)
}
\examples{
test = dplyr::tibble(date = as.Date("2017-01-01") + 0:9, value = rnorm(10))
arima11_fun(test, 1)
}
