% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_forecastors.R
\name{arima_estimator}
\alias{arima_estimator}
\title{arima estimator}
\usage{
arima_estimator(train_sample, predicted, options, predictors = "")
}
\arguments{
\item{train_sample}{tsibble with train sample}

\item{predicted}{name of the predicted variable}

\item{options}{character with options separated by comma
Available options are: p, d, q, pseas, dseas, qseas, method.}

\item{predictors}{character with predictors separated by plus or comma, no predictors by default}
}
\value{
fitted arima model
}
\description{
Envelope function that estimates arima model
}
\details{
Envelope function that estimates arima model. If no p, d, q are specified then auto.arima is fitted.
}
\examples{
train_sample = tsibble::tsibble(date = lubridate::ymd("2017-01-01") + months(0:99),
    y = arima.sim(n = 100, model = list(ar = 0.7)), x = 1:100)
arima_estimator(train_sample, "y", "p=2,d=0,q=0", "x")
}
