% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_forecastors.R
\name{arima_forecastor}
\alias{arima_forecastor}
\title{arima forecastor}
\usage{
arima_forecastor(
  fit,
  test_sample = NA,
  predicted = "",
  predicted_ = "",
  options = "",
  predictors = "",
  h = 1,
  frequency = 1
)
}
\arguments{
\item{fit}{fitted arima model}

\item{test_sample}{tsibble with predictors for test sample}

\item{predicted}{name of all the predicted variables. Not used as is determined by the fit.}

\item{predicted_}{name of the perdicted variable. Not used as is determined by the fit.}

\item{options}{character with options separated by comma
not supparted}

\item{predictors}{character with predictors separated by plus or comma, no predictors by default}

\item{h}{forecasting horizon}

\item{frequency}{time series frequency
may be it is needed for forecasting with regressors, may check later}
}
\value{
forecast object
}
\description{
Envelope function that forecasts arima model
}
\details{
Envelope function that forecasts arima model.
}
\examples{
train_sample = tsibble::tsibble(date = lubridate::ymd("2017-01-01") + months(0:99),
    y = arima.sim(n = 100, model = list(ar = 0.7)), x = 1:100)
fit = arima_estimator(train_sample, "y", "p=2,d=0,q=0", "x")
test_sample = tsibble::tsibble(date = lubridate::ymd("2030-01-01") + months(0:1), x = 5:6)
arima_forecastor(fit, test_sample, predictors = "x", h = 2, frequency = 12)
}
