% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{arima_fun}
\alias{arima_fun}
\title{Do forecast using auto ARIMA}
\usage{
arima_fun(model_sample, h)
}
\arguments{
\item{model_sample}{preferably tsibble with "value" column}

\item{h}{forecasting horizon, is ignored}
}
\value{
auto ARIMA model
}
\description{
Do forecast using auto ARIMA
}
\details{
Do forecast using auto ARIMA
}
\examples{
test = dplyr::tibble(date = as.Date("2017-01-01") + 0:9, value = rnorm(10))
arima_fun(test, 1)
}
