% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_operations.R
\name{as_numeric_cyrillic}
\alias{as_numeric_cyrillic}
\title{Convert russian numbers with comma into numeric}
\usage{
as_numeric_cyrillic(x)
}
\arguments{
\item{x}{vector of russian numbers, i.e. "45,34"}
}
\value{
numeric vectors
}
\description{
Convert russian numbers with comma into numeric
}
\details{
Converts russian numbers with comma into numeric. 
Comma is replaced by dot and than `as.numeric` is applied.
Spaces are removed!
}
\examples{
as_numeric_cyrillic(c("56,23", "-23,12"))
}
