% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{augment_tsibble_4_regression}
\alias{augment_tsibble_4_regression}
\title{Augment tsibble with usual regressors}
\usage{
augment_tsibble_4_regression(original, dependent = "value", h = 1)
}
\arguments{
\item{original}{original tsibble}

\item{dependent}{name of the dependent variable}

\item{h}{number of rows to append in the future}
}
\value{
augmented tibble
}
\description{
Augment tsibble with usual regressors
}
\details{
Augment tsibble with usual regressors.
Adds trend, fourier terms, lags of regressor and dependen variables.
Also appends h rows for forecasting.
}
\examples{
test_ts = stats::ts(rnorm(100), start = c(2000, 1), freq = 12)
test_tsibble = tsibble::as_tsibble(test_ts)
augment_tsibble_4_regression(test_tsibble, h = 4)
}
