% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{calculate_mae_table}
\alias{calculate_mae_table}
\title{Calculate mae table from estimated models tibble}
\usage{
calculate_mae_table(model_list_fitted)
}
\arguments{
\item{model_list_fitted}{tibble with one model per row}
}
\value{
tibble with mae
}
\description{
Calculate mae table from estimated models tibble
}
\details{
Calculate mae table from estimated models tibble.
In this tibble dependent variable is always named value.
}
\examples{
# no yet
}
