% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{check_conversion}
\alias{check_conversion}
\title{Checks conversion results}
\usage{
check_conversion(data_tsibble)
}
\arguments{
\item{data_tsibble}{converted time series}
}
\value{
TRUE or will stop with error message
}
\description{
Checks some basic requirements for converted data from raw formats.
}
\details{
Checks some basic requirements for converted data from raw formats.
}
\examples{
\donttest{
cpi = convert_i_ipc_xlsx()
check_conversion(cpi)
}
}
