% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_1_06_0_xlsx}
\alias{convert_1_06_0_xlsx}
\title{Converts Investment file from rosstat to tibble}
\usage{
convert_1_06_0_xlsx(
  path_to_source = "http://www.gks.ru/bgd/regl/b19_02/IssWWW.exe/Stg/d010/1-06-0.xlsx",
  access_date = Sys.Date()
)
}
\arguments{
\item{path_to_source}{name of the original 1-06-0.xls file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts Investment file from rosstat to tibble
}
\details{
Converts Investment file from rosstat to tibble
Written by Rifat Eniliev
}
\examples{
\donttest{
invest = convert_invest_xlsx()
}
}
