% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_1_nn_doc}
\alias{convert_1_nn_doc}
\title{Converts 1-nn file from rosstat to tibble}
\usage{
convert_1_nn_doc(
  path_to_source = "http://www.gks.ru/bgd/regl/b18_02/IssWWW.exe/Stg/d010/1-08.doc",
  access_date = Sys.Date()
)
}
\arguments{
\item{path_to_source}{name of the original 1-nn.doc file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts 1-nn file from rosstat to tibble
}
\details{
Converts 1-nn file from rosstat to tibble.
Written by: Rifat Enileev
}
\examples{
\donttest{
one = convert_1_nn_doc()
one = convert_1_nn_doc("http://www.gks.ru/bgd/regl/b18_02/IssWWW.exe/Stg/d010/1-08.doc")
two = convert_1_nn_doc("http://www.gks.ru/bgd/regl/b18_02/IssWWW.exe/Stg/d010/1-03.doc")
three = convert_1_nn_doc("http://www.gks.ru/bgd/regl/b18_02/IssWWW.exe/Stg/d010/1-11.doc")
}
}
