% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_1_nn_xlsx}
\alias{convert_1_nn_xlsx}
\title{Converts 1-nn xlsx file from rosstat to tibble}
\usage{
convert_1_nn_xlsx(
  path_to_source = "https://www.gks.ru/bgd/regl/b20_02/IssWWW.exe/Stg/d010/1-11.xlsx",
  access_date = Sys.Date()
)
}
\arguments{
\item{path_to_source}{name of the original 1-nn.doc file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts 1-nn xlsx file from rosstat to tibble
}
\details{
Converts 1-nn xlsx file from rosstat to tibble.
Written by: Rifat Enileev
}
\examples{
\donttest{
one = convert_1_nn_xlsx("https://www.gks.ru/bgd/regl/b20_02/IssWWW.exe/Stg/d010/1-07.xlsx")
two = convert_1_nn_xlsx("https://www.gks.ru/bgd/regl/b20_02/IssWWW.exe/Stg/d010/1-03.xlsx")
three = convert_1_nn_xlsx("https://www.gks.ru/bgd/regl/b20_02/IssWWW.exe/Stg/d010/1-11.xlsx")
}
}
