% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_ind_baza_2018_xlsx}
\alias{convert_ind_baza_2018_xlsx}
\title{Converts ind_baza_2018 file from rosstat to tibble}
\usage{
convert_ind_baza_2018_xlsx(
 
    path_to_source = "https://rosstat.gov.ru/storage/mediabank/YMKvI51h/ind_baza-2018.xlsx",
  access_date = Sys.Date()
)
}
\arguments{
\item{path_to_source}{name of the original ind-baza-2018.xlsx file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts ind_baza_2018 file from rosstat to tibble
}
\details{
Converts ind_baza_2018 file from rosstat to tibble.
Written by: Nastya Jarkova, mod by Rifat Enileev
}
\examples{
\donttest{
# ind = convert_ind_baza_2018_xlsx()
# removed by gks 2020-09-20
# Old link https://gks.ru/storage/mediabank/ind-baza-2018.xlsx
# New working link https://rosstat.gov.ru/storage/mediabank/YMKvI51h/ind_baza-2018.xlsx
}
}
