% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_ind_okved2_xls}
\alias{convert_ind_okved2_xls}
\title{Converts Ind_sub-2018.xls file from rosstat to tibble}
\usage{
convert_ind_okved2_xls(
  path_to_source = "https://rosstat.gov.ru/storage/mediabank/BYkjy3Bn/Ind_sub-2018.xls",
  access_date = Sys.Date()
)
}
\arguments{
\item{path_to_source}{name of the original ind_okved2.xlsx file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Written by: Vladimir Omelyusik
}
\examples{
\donttest{
# ind = convert_ind_okved2_xls()
# new link https://rosstat.gov.ru/storage/mediabank/BYkjy3Bn/Ind_sub-2018.xls
}

}
