% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_ind_okved2_xls_one_sheet}
\alias{convert_ind_okved2_xls_one_sheet}
\title{Converts one sheet of ind-baza-2018.xlsx or Ind_sub-2018.xls file from rosstat to tibble}
\usage{
convert_ind_okved2_xls_one_sheet(
  path_to_source = "https://rosstat.gov.ru/storage/mediabank/BYkjy3Bn/Ind_sub-2018.xls",
  access_date = Sys.Date(),
  sheet = 1,
  variable_name = "ind_prod"
)
}
\arguments{
\item{path_to_source}{name of the original ind_okved2.xlsx file}

\item{access_date}{date of access is appended to every observation}

\item{sheet}{number of sheet}
}
\value{
tibble
}
\description{
Written by: Vladimir Omelyusik
}
\examples{
\donttest{
# ind = convert_ind_okved2_xls()
# new link https://rosstat.gov.ru/storage/mediabank/BYkjy3Bn/Ind_sub-2018.xls
}

}
