% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_ind_okved2_xlsx}
\alias{convert_ind_okved2_xlsx}
\title{Converts ind_okved2 or ind_baza_2018 file from rosstat to tibble}
\usage{
convert_ind_okved2_xlsx(
  path_to_source = "http://www.gks.ru/free_doc/new_site/business/prom/ind_okved2.xlsx",
  access_date = Sys.Date()
)
}
\arguments{
\item{path_to_source}{name of the original ind_okved2.xlsx file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts ind_okved2 or ind_baza_2018 file from rosstat to tibble
}
\details{
Converts ind_okved2 or ind_baza_2018 file from rosstat to tibble.
Written by: Nastya Jarkova, Rifat Enileev, etc
}
\examples{
\donttest{
# ind = convert_ind_okved2_xlsx()
# old link http://www.gks.ru/storage/mediabank/ind_okved2(1).xlsx
# old link https://gks.ru/storage/mediabank/ind-baza-2018.xlsx
# old link https://rosstat.gov.ru/storage/mediabank/YMKvI51h/ind_baza-2018.xlsx
# new http://www.gks.ru/free_doc/new_site/business/prom/ind_okved2.xlsx
# new https://rosstat.gov.ru/storage/mediabank/6CWGi6WR/ind_baza-2018.xlsx
}

}
