% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_lendrate}
\alias{convert_lendrate}
\title{Parse lending rates from cbr}
\usage{
convert_lendrate(
  path_to_source = "http://www.cbr.ru/hd_base/mkr/mkr_monthes/",
  access_date = Sys.Date()
)
}
\arguments{
\item{path_to_source}{name of the original html}

\item{access_date}{date of access is appended to every observation}
}
\value{
tsibble
}
\description{
Parse lending rates from cbr
}
\details{
Parse lending rates from cbr
Written by Nastya Jarkova
}
\examples{
\donttest{
lend_rate = convert_lendrate()
}
}
