% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_m2_m2_sa}
\alias{convert_m2_m2_sa}
\title{Converts m2-m2_sa file from rosstat to tibble}
\usage{
convert_m2_m2_sa(path_to_source, access_date = Sys.Date())
}
\arguments{
\item{path_to_source}{name of the original m2-m2_sa.docx file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts m2-m2_sa file from rosstat to tibble
}
\details{
Converts m2-m2_sa file from rosstat to tibble.
Written by: Petr Garmider
}
\examples{
# no yet
}
