% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_m2_m2_sa_xlsx}
\alias{convert_m2_m2_sa_xlsx}
\title{Converts m2-m2_sa file from rosstat to tibble}
\usage{
convert_m2_m2_sa_xlsx(
  path_to_source = "http://www.cbr.ru/vfs/statistics/credit_statistics/M2-M2_SA.xlsx",
  access_date = Sys.Date()
)
}
\arguments{
\item{path_to_source}{name of the original m2-m2_sa.xlsx file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts m2-m2_sa file from rosstat to tibble
}
\details{
Converts m2-m2_sa file from rosstat to tibble.
Written by: Petr Garmider
}
\examples{
\donttest{
m2sa = convert_m2_m2_sa_xlsx()
}
}
