% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_reserves}
\alias{convert_reserves}
\title{Parse reserves data from cbr}
\usage{
convert_reserves(
  path_to_source = "http://www.cbr.ru/hd_base/mrrf/mrrf_m/",
  access_date = Sys.Date()
)
}
\arguments{
\item{path_to_source}{name of the original html}

\item{access_date}{date of access is appended to every observation}
}
\value{
tsibble
}
\description{
Parse reserves data from cbr
}
\details{
Parse reserves data from cbr
Written by Petr Garmider
}
\examples{
\donttest{
reserves <- convert_reserves()
}
}
