% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_tab5a_xls}
\alias{convert_tab5a_xls}
\title{Converts tab5a file from rosstat to tibble}
\usage{

  convert_tab5a_xls(path_to_source = "http://www.gks.ru/free_doc/new_site/vvp/kv/tab5a.xls",
  access_date = Sys.Date())
}
\arguments{
\item{path_to_source}{name of the original tab5a.xls file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts tab5a file from rosstat to tibble
}
\details{
Converts tab5a file from rosstat to tibble.
The structure is similar to tab6b.
}
\examples{
\donttest{
tab5a = convert_tab5a_xls()
}
}
