% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_tab6b_xls}
\alias{convert_tab6b_xls}
\title{Converts tab6b file from rosstat to tibble}
\usage{
convert_tab6b_xls(
  path_to_source = "http://www.gks.ru/free_doc/new_site/vvp/kv/tab6b.xls",
  access_date = Sys.Date()
)
}
\arguments{
\item{path_to_source}{name of the original tab6b.xls file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts tab6b file from rosstat to tibble
}
\details{
Converts tab6b file from rosstat to tibble.
The structure is similar to tab5a.
}
\examples{
\donttest{
tab6b = convert_tab6b_xls()
}
}
