% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_tab9_xls}
\alias{convert_tab9_xls}
\title{Converts tab9 file from rosstat to tibble}
\usage{

  convert_tab9_xls(path_to_source = "http://www.gks.ru/free_doc/new_site/vvp/kv/tab9.xls",
  access_date = Sys.Date())
}
\arguments{
\item{path_to_source}{name of the original tab9.xls file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts tab9 file from rosstat to tibble
}
\details{
Converts tab9 file from rosstat to tibble.
}
\examples{
\donttest{
tab9 = convert_tab9_xls()
}
}
