% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_trade_xls}
\alias{convert_trade_xls}
\title{Converts trade.xls file from cbr to tibble}
\usage{
convert_trade_xls(
 
    path_to_source = "https://www.cbr.ru/vfs/statistics/credit_statistics/trade/trade.xls",
  access_date = Sys.Date()
)
}
\arguments{
\item{path_to_source}{name of the original trade.xls file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts trade.xls file from cbr to tibble
}
\details{
Converts trade.xls file from cbr to tibble.
Written by: Maxim Alekseev
}
\examples{
\donttest{
trade = convert_trade_xls()
}
}
