% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_string.R
\name{date_to_string}
\alias{date_to_string}
\title{Converts strings into dates in Russian language}
\usage{
date_to_string(date, freq = 12)
}
\arguments{
\item{date}{date in the following format: "2019-02-23"}

\item{freq}{equal to 4 for quarters, equal to 12 for months}
}
\value{
date in the form of the string
}
\description{
Converts strings into dates in Russian language
}
\details{
Converts strings into dates in Russian language
}
\examples{
date <- c("2019-02-23", "2011-01-07")
date_to_string(date, freq = 12)
}
