% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchdog.R
\name{download_statistics}
\alias{download_statistics}
\title{Download all statistics}
\usage{
download_statistics(path, watchdog, access_date = Sys.Date())
}
\arguments{
\item{path}{path to raw data folde}

\item{watchdog}{watchdog file}

\item{access_date}{access date}
}
\value{
downloads log
}
\description{
Download all statistics
}
\details{
Download all statistics
}
\examples{
path = "../raw/"
# download_log_new = download_statistics(path, watchdog)
}
