% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchdog.R
\name{download_watchdog_files}
\alias{download_watchdog_files}
\title{Download files guarded by watchdog}
\usage{
download_watchdog_files(raw_data_folder, watchdog_file = "watchdog.csv")
}
\arguments{
\item{raw_data_folder}{path to raw data folde}

\item{watchdog_file}{name of watchdog file}
}
\value{
updated version of watchdog file
}
\description{
Download files guarded by watchdog
}
\details{
Download files guarded by watchdog
}
\examples{
raw_data_folder = "../raw/"
# new_watchdog = download_watchdog_files(raw_data_folder)
}
