% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{estimate_and_forecast}
\alias{estimate_and_forecast}
\title{Estimate and forecast all models from model tibble}
\usage{
estimate_and_forecast(model_list)
}
\arguments{
\item{model_list}{tibble with one model per row}
}
\value{
tibble with estimated models and point forecasts
}
\description{
Estimate and forecast all models from model tibble
}
\details{
Estimate and forecast all models from model tibble
In this tibble dependent variable is always named value.
}
\examples{
# no yet
}
