% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{estimate_nonduplicate_models}
\alias{estimate_nonduplicate_models}
\title{Estimate non-duplicate models from model tibble}
\usage{
estimate_nonduplicate_models(model_list, store_models = c("tibble",
  "file"))
}
\arguments{
\item{model_list}{tibble with one model per row}

\item{store_models}{inside tibble or in one separate file per model}
}
\value{
rows of the orignal tibble correspinding to non-duplicate models plus column with estimated models
}
\description{
Estimate non-duplicate models from model tibble
}
\details{
Estimate non-duplicate models from model tibble.
If model is the same for different h it is estimated only once.
}
\examples{
# no yet
}
