% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_forecastors.R
\name{ets_forecastor}
\alias{ets_forecastor}
\title{ets forecastor}
\usage{
ets_forecastor(
  fit,
  test_sample = NA,
  predicted = "",
  predicted_ = "",
  options = "",
  predictors = "",
  h = 1,
  frequency = 1
)
}
\arguments{
\item{fit}{fitted ets model}

\item{test_sample}{not used by ets model}

\item{predicted}{name of all the predicted variables. Not used as is determined by the fit.}

\item{predicted_}{name of the perdicted variable. Not used as is determined by the fit.}

\item{options}{character with options separated by comma
not supparted}

\item{predictors}{not supported by ets model}

\item{h}{forecasting horizon}

\item{frequency}{time series frequency. not used by ets model}
}
\value{
forecast object
}
\description{
Envelope function that forecasts ets model
}
\details{
Envelope function that forecasts ets model.
}
\examples{
train_sample = tsibble::tsibble(date = lubridate::ymd("2017-01-01") + months(0:99),
    y = arima.sim(n = 100, model = list(ar = 0.7)), x = 1:100)
fit = arima_estimator(train_sample, "y", "p=2,d=0,q=0")
ets_forecastor(fit, h = 2)
}
