% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{ets_fun}
\alias{ets_fun}
\title{Do forecast using auto ETS}
\usage{
ets_fun(model_sample, h, target = "value")
}
\arguments{
\item{model_sample}{preferably tsibble with "value" column}

\item{h}{forecasting horizon, is ignored}

\item{target}{name of the target variable, "value" by default}
}
\value{
auto ETS model
}
\description{
Do forecast using auto ETS
}
\details{
Do forecast using auto ETS
}
\examples{
test = dplyr::tibble(date = as.Date("2017-01-01") + 0:9, value = rnorm(10))
ets_fun(test, 1)
}
