% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchdog.R
\name{export_with_safe_date}
\alias{export_with_safe_date}
\title{Exports tsibble to csv correctly}
\usage{
export_with_safe_date(x, file)
}
\arguments{
\item{x}{tibble or tsibble with date column}

\item{file}{filename}
}
\value{
file exporting code
}
\description{
Exports tsibble to csv correctly
}
\details{
Exports tsibble to csv correctly
The problem is that simple rio::export will transform monthly dates into strange integers.
}
\examples{
x = ts(1:10, frequency = 12, start = c(2000, 1))
x_tsi = tsibble::as_tsibble(x)
x_tsi = dplyr::rename(x_tsi, date = index)
export_with_safe_date(x_tsi, paste0(tempdir(), "/data.csv"))
}
