% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_value}
\alias{extract_value}
\title{Extract value column and present it as ts}
\usage{
extract_value(model_sample)
}
\arguments{
\item{model_sample}{preferably tsibble with "value" column}
}
\value{
univariate time series
}
\description{
Extract value column and present it as ts
}
\details{
Extract value column and present it as ts
}
\examples{
test = tsibble(date = as.Date("2017-01-01") + 0:9, value = rnorm(10))
extract_value(test)
}
