% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fill_duplicate_models}
\alias{fill_duplicate_models}
\title{Fill duplicate models into model tibble}
\usage{
fill_duplicate_models(model_list_half_fitted, full_model_list)
}
\arguments{
\item{model_list_half_fitted}{tibble with estimated non-duplicate models}

\item{full_model_list}{tibble with complete list of duplicated and non-duplicated models}
}
\value{
full original tibble with one estimated model per row
}
\description{
Fill duplicate models into model tibble
}
\details{
Fill duplicate models into model tibble.
}
\examples{
# no yet
}
