% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast_2_scalar}
\alias{forecast_2_scalar}
\title{Extract one scalar forecast from forecast object}
\usage{
forecast_2_scalar(forecast_object, h = 1)
}
\arguments{
\item{forecast_object}{forecast object}

\item{h}{forecasting horizon}
}
\value{
mean scalar forecast
}
\description{
Extract one scalar forecast from forecast object
}
\details{
Extract one scalar forecast from forecast object
}
\examples{
test = dplyr::tibble(date = as.Date("2017-01-01") + 0:9, value = rnorm(10))
tbats = tbats_fun(test, 1)
fcst = forecast(tbats, h = 2)
forecast_2_scalar(fcst, h = 2)
}
