% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_processing.R
\name{gentle_as_numeric}
\alias{gentle_as_numeric}
\title{transforms all numeric strings to numbers keeping non-numeric}
\usage{
gentle_as_numeric(chr_vector)
}
\arguments{
\item{chr_vector}{character vector or list}
}
\value{
list with numerics and character 1d objects
}
\description{
transforms all numeric strings to numbers keeping non-numeric
}
\details{
transform all numeric strings to numbers keeping non-numeric, returns a list.
}
\examples{
gentle_as_numeric(c("5", "ggg", "'sss'"))
}
