% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_first_date}
\alias{get_first_date}
\title{Get first date from tsibble}
\usage{
get_first_date(original)
}
\arguments{
\item{original}{original tsibble}
}
\value{
first date
}
\description{
Get first date from tsibble
}
\details{
Get first date from tsibble
}
\examples{
test_ts = stats::ts(rnorm(100), start = c(2000, 1), freq = 12)
test_tsibble = tsibble::as_tsibble(test_ts)
get_first_date(test_tsibble)
}
