% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_last_date}
\alias{get_last_date}
\title{Get last date from tsibble}
\usage{
get_last_date(original)
}
\arguments{
\item{original}{original tsibble}
}
\value{
last date
}
\description{
Get last date from tsibble
}
\details{
Get last date from tsibble
}
\examples{
test_ts = stats::ts(rnorm(100), start = c(2000, 1), freq = 12)
test_tsibble = tsibble::as_tsibble(test_ts)
get_last_date(test_tsibble)
}
