% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_processing.R
\name{get_last_n_obs}
\alias{get_last_n_obs}
\title{Get last high frequency observations}
\usage{
get_last_n_obs(hf_data, lf_date, n_lags = 1, one_row = TRUE)
}
\arguments{
\item{hf_data}{high frequency data}

\item{lf_date}{low frequency date}

\item{n_lags}{number of lags to add}

\item{one_row}{output representation}
}
\value{
tsibble with high frequency lags added
}
\description{
Get last high frequency observations
}
\details{
Get last high frequency observations
}
\examples{
# no yet
}
