% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchdog.R
\name{get_last_version_download_date}
\alias{get_last_version_download_date}
\title{Gets last version date from watchdog file}
\usage{
get_last_version_download_date(source_url, watchdog)
}
\arguments{
\item{source_url}{url of guarded file}

\item{watchdog}{data frame with url and last_download columns at least}
}
\value{
date of last downloaded version
}
\description{
Gets last version date from watchdog file
}
\details{
Gets last version date from watchdog file
}
\examples{
watchdog_demo = dplyr::tibble(url = c("a", "b"), file = c("xxx.xls", "yyy.xlsx"), 
  last_download = c("2011-11-11", "2010-10-10"))
get_last_version_download_date("a", watchdog_demo)
}
