% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchdog.R
\name{get_last_version_path}
\alias{get_last_version_path}
\title{Gets last version full path from watchdog file}
\usage{
get_last_version_path(source_url, watchdog)
}
\arguments{
\item{source_url}{url of guarded file}

\item{watchdog}{data frame with url, file and last_download columns at least}
}
\value{
path to last downloaded version
}
\description{
Gets last version full path from watchdog file
}
\details{
Gets last version full path from watchdog file
}
\examples{
watchdog_demo = dplyr::tibble(url = c("a", "b"), file = c("xxx.xls", "yyy.xlsx"), last_download = c("2011-11-11", "2010-10-10"))
get_last_version_path("a", watchdog_demo)
}
