% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchdog.R
\name{get_watchdog_line}
\alias{get_watchdog_line}
\title{Extracts one line from watchdog file given url}
\usage{
get_watchdog_line(source_url, watchdog)
}
\arguments{
\item{source_url}{url of guarded file}

\item{watchdog}{data frame with url column at least}
}
\value{
one line from watchdog data frame
}
\description{
Extracts one line from watchdog file given url
}
\details{
Extracts one line from watchdog file given url
}
\examples{
watchdog_demo = dplyr::tibble(url = c("a", "b"), x = c(1, 2))
get_watchdog_line("a", watchdog_demo)
}
