% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{I_ipc_xls_convert}
\alias{I_ipc_xls_convert}
\title{Converts I_ipc file from rosstat to tibble}
\usage{
I_ipc_xls_convert(path_to_source, access_date)
}
\arguments{
\item{path_to_source}{name of the original I_ipc.xls file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts I_ipc file from rosstat to tibble
}
\details{
Converts I_ipc file from rosstat to tibble
}
\examples{
# no yet
}
