% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{lasso_2_scalar_forecast}
\alias{lasso_2_scalar_forecast}
\title{Obtain scalar forecast from lasso model}
\usage{
lasso_2_scalar_forecast(model, h = 1, dependent = "value",
  model_sample, s = c("lambda.min", "lambda.1se"))
}
\arguments{
\item{model}{estimated lasso model}

\item{h}{forecasting horizon}

\item{dependent}{name of the dependent variable}

\item{model_sample}{non-augmented data set for model estimation}

\item{s}{criterion to select best regularization lambda in lasso}
}
\value{
scalar forecast for given h
}
\description{
Obtain scalar forecast from lasso model
}
\details{
Obtain scalar forecast from lasso model.
The function automatically augments data with lags, fourier terms, trend etc.
}
\examples{
test_ts = stats::ts(rnorm(100), start = c(2000, 1), freq = 12)
test_tsibble = tsibble::as_tsibble(test_ts)
model = lasso_fun(test_tsibble, h = 1)
lasso_2_scalar_forecast(model, h = 1, model_sample = test_tsibble)
}
