% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{lasso_augmented_estimate}
\alias{lasso_augmented_estimate}
\title{Estimate lasso model using tsibble with regressors}
\usage{
lasso_augmented_estimate(augmented, target = "value", seed = 777)
}
\arguments{
\item{augmented}{tsibble with all predictors with lags.
May be obtained using `augment_tsibble_4_regression`.}

\item{target}{name of the target variable}

\item{seed}{random seed}
}
\value{
lasso model
}
\description{
Estimate lasso model using tsibble with regressors
}
\details{
Estimate lasso model using tsibble with regressors
Regressors should already include lags, fourier terms, trend etc
}
\examples{
test_ts = stats::ts(rnorm(100), start = c(2000, 1), freq = 12)
test_tsibble = tsibble::as_tsibble(test_ts) \%>\% dplyr::rename(date = index)
augmented = augment_tsibble_4_regression(test_tsibble, h = 4)
model = lasso_augmented_estimate(augmented)
}
