% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{lasso_fun}
\alias{lasso_fun}
\title{Augment data and estimate lasso model}
\usage{
lasso_fun(model_sample, seed = 777, target = "value", h = 1)
}
\arguments{
\item{model_sample}{tsibble that will be augmented with trend etc}

\item{seed}{random seed}

\item{target}{name of the target variable}

\item{h}{forecasting horizon}
}
\value{
lasso model
}
\description{
Augment data and estimate lasso model
}
\details{
Augment data and estimate lasso model.
Trend, fourier terms, lags are added before estimation of lasso model.
}
\examples{
test_ts = stats::ts(rnorm(100), start = c(2000, 1), freq = 12)
test_tsibble = tsibble::as_tsibble(test_ts)
model = lasso_fun(test_tsibble)
}
