% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{month_to_genetivus}
\alias{month_to_genetivus}
\title{Changes month name from nominativus to genetivus}
\usage{
month_to_genetivus(x)
}
\arguments{
\item{x}{string with month names in nominativus}
}
\value{
string with month names in nominativus
}
\description{
Changes month name from nominativus to genetivus.
}
\details{
Functions like dmy needs month names in genetivus.
}
\examples{
month_to_genetivus("май")
}
