% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_processing.R
\name{numbers_string_2_vector}
\alias{numbers_string_2_vector}
\title{converts string with comma separated numbers to numeric vector}
\usage{
numbers_string_2_vector(numbers_string)
}
\arguments{
\item{numbers_string}{character, something like '1,2,3,7.65'}
}
\value{
numeric vector
}
\description{
converts string with comma separated numbers to numeric vector
}
\details{
converts string with comma separated numbers to numeric vector.
}
\examples{
numbers_string_2_vector("1,2,4.56")
}
