% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_processing.R
\name{param_string_2_tibble}
\alias{param_string_2_tibble}
\title{converts string with parameter values to tibble}
\usage{
param_string_2_tibble(param_string)
}
\arguments{
\item{param_string}{character}
}
\value{
tibble with columns names equal to parameter names
}
\description{
converts string with parameter values to tibble
}
\details{
Converts string with parameter values to tibble. 
Parameters should be separated by comma.
}
\examples{
param_string_2_tibble("q=3, p=4, v='ml', qur=mlp")
}
