% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_statistics.R
\name{parse_lendrate}
\alias{parse_lendrate}
\title{Parse lending rates from cbr}
\usage{
parse_lendrate(access_date = Sys.Date())
}
\arguments{
\item{access_date}{date of access is appended to every observation}
}
\value{
tsibble
}
\description{
Parse lending rates from cbr
}
\details{
Parse lending rates from cbr
Written by Nastya Jarkova
}
\examples{
# lend_rate <- parse_lendrate()
}
