% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_statistics.R
\name{parse_reserverate}
\alias{parse_reserverate}
\title{Parse reserve rates from cbr}
\usage{
parse_reserverate(access_date = Sys.Date())
}
\arguments{
\item{access_date}{date of access is appended to every observation}
}
\value{
tsibble
}
\description{
Parse reserve rates from cbr
}
\details{
Parse reserve rates from cbr
Written by Petr Garmider
}
\examples{
res_rate <- parse_reserverate()
}
