% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_statistics.R
\name{parse_reserves}
\alias{parse_reserves}
\title{Parse reserves data from cbr}
\usage{
parse_reserves(access_date = Sys.Date())
}
\arguments{
\item{access_date}{date of access is appended to every observation}
}
\value{
tsibble
}
\description{
Parse reserves data from cbr
}
\details{
Parse reserves data from cbr
Written by Petr Garmider
}
\examples{
# res_rate <- parse_reserverate()
}
