% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{prepare_model_list}
\alias{prepare_model_list}
\title{Prepare model tibble for cross-validation}
\usage{
prepare_model_list(
  h_all = 1,
  model_fun_tibble,
  series_data,
  dates_test,
  window_type = c("sliding", "stretching"),
  target = "value"
)
}
\arguments{
\item{h_all}{vector of forecasting horizons}

\item{model_fun_tibble}{tibble with names of estimator functions and scalar forecast extractors}

\item{series_data}{tsibble with full data sample}

\item{dates_test}{test dates}

\item{window_type}{sliding or stretching}

\item{target}{name of the target variable}
}
\value{
tibble with one row per model
}
\description{
Prepare model tibble for cross-validation
}
\details{
Prepare model tibble for cross-validation
}
\examples{
# no yet
}
