% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{prepare_model_list2}
\alias{prepare_model_list2}
\title{Prepare model tibble for forecasts}
\usage{
prepare_model_list2(h_all = 1, model_fun_tibble, series_data,
  dependent = "value")
}
\arguments{
\item{h_all}{vector of forecasting horizons}

\item{model_fun_tibble}{tibble with names of estimator functions and scalar forecast extractors}

\item{series_data}{tsibble with full data sample}

\item{dependent}{name of the dependent variable}
}
\value{
tibble with one row per model
}
\description{
Prepare model tibble for forecasts
}
\details{
Prepare model tibble for forecasts
}
\examples{
# no yet
}
