% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{raise_warnings}
\alias{raise_warnings}
\title{Produces warnings table}
\usage{
raise_warnings(report_table)
}
\arguments{
\item{report_table}{downloaded series' metadata}
}
\value{
data.frame
}
\description{
Raises warnings and produces a table of suspicious metrics' values.
}
\details{
Use warnings() to see them.
}
\examples{
\donttest{
report_table <- make_report(download_log_new, path_day)
warnings_table <- raise_warnings(report_table)
}
}
