% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{ranger_2_scalar_forecast}
\alias{ranger_2_scalar_forecast}
\title{Obtain scalar forecast from random forest (ranger) model}
\usage{
ranger_2_scalar_forecast(model, h = 1, target = "value", model_sample)
}
\arguments{
\item{model}{estimated ranger model}

\item{h}{forecasting horizon}

\item{target}{name of the target variable}

\item{model_sample}{non-augmented data set for model estimation}
}
\value{
scalar forecast for given h
}
\description{
Obtain scalar forecast from random forest (ranger) model
}
\details{
Obtain scalar forecast from random forest (ranger) model.
The function automatically augments data with lags, fourier terms, trend etc.
}
\examples{
test_ts = stats::ts(rnorm(100), start = c(2000, 1), freq = 12)
test_tsibble = tsibble::as_tsibble(test_ts)
model = ranger_fun(test_tsibble, h = 1)
ranger_2_scalar_forecast(model, h = 1, model_sample = test_tsibble)
}
