% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{ranger_fun}
\alias{ranger_fun}
\title{Augment data and estimate ranger model}
\usage{
ranger_fun(model_sample, seed = 777, target = "value", h = 1)
}
\arguments{
\item{model_sample}{tsibble that will be augmented with trend etc}

\item{seed}{random seed}

\item{target}{name of the target variable}

\item{h}{forecasting horizon}
}
\value{
ranger model
}
\description{
Augment data and estimate ranger model
}
\details{
Augment data and estimate random forest (ranger) model.
Trend, fourier terms, lags are added before estimation of lasso model.
}
\examples{
test_ts = stats::ts(rnorm(100), start = c(2000, 1), freq = 12)
test_tsibble = tsibble::as_tsibble(test_ts)
model = ranger_fun(test_tsibble)
}
