% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_operations.R
\name{read_datastream}
\alias{read_datastream}
\title{Read excel files saved from data stream}
\usage{
read_datastream(filename, correct_names = TRUE)
}
\arguments{
\item{filename}{name of the excel file from datastream}

\item{correct_names}{logical (TRUE) whether column names should be corrected}
}
\value{
data frame
}
\description{
Read excel files saved from data stream
}
\details{
Read excel files saved from data stream
Transform bad column names
}
