% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_processing.R
\name{remove_quotes}
\alias{remove_quotes}
\title{removes quotes around character vectors}
\usage{
remove_quotes(chr_vector)
}
\arguments{
\item{chr_vector}{character vector or list}
}
\value{
character vector with quotes removed
}
\description{
removes quotes around character vectors
}
\details{
removes quotes around character vectors.
}
\examples{
remove_quotes(c("aaa", "'bbb'"))
}
