% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchdog.R
\name{replace_extension}
\alias{replace_extension}
\title{Replace file extension by specified ending}
\usage{
replace_extension(filename, new_ext = "_converted.csv")
}
\arguments{
\item{filename}{name of file with extension}

\item{new_ext}{string, new ending of full file name}
}
\value{
new file name
}
\description{
Replace file extension by specified ending
}
\details{
Replace file extension by specified ending
Old extension is removed and "_converted.csv" is added
}
\examples{
replace_extension("../test/demo.xlsx")
}
