% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_processing.R
\name{split_variable_names}
\alias{split_variable_names}
\title{splits comma or plus separated variable names into a character vector}
\usage{
split_variable_names(predictors_vector, acronyms = NULL, split_by = "[\\\\+,]")
}
\arguments{
\item{predictors_vector}{character vector of comma or plus separated variable lists}

\item{acronyms}{tibble with two character columns: acronym, meaning}

\item{split_by}{regex expression, comma and plus by default}
}
\value{
character vector with variable names
}
\description{
splits comma or plus separated variable names into a character vector
}
\details{
splits comma or plus separated variable names into a character vector. 
Duplicate variable names are removed. For a vector input just collects all variable names.
}
\examples{
split_variable_names(c("aaa+bbb", "ccc,bbb,ddd"))
}
