% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{tab5a_xls_convert}
\alias{tab5a_xls_convert}
\title{Converts tab5a file from rosstat to tibble}
\usage{
tab5a_xls_convert(path_to_source, access_date)
}
\arguments{
\item{path_to_source}{name of the original tab5a.xls file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts tab5a file from rosstat to tibble
}
\details{
Converts tab5a file from rosstat to tibble.
The structure is similar to tab6b.
}
\examples{
# no yet
}
