% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{tab6b_xls_convert}
\alias{tab6b_xls_convert}
\title{Converts tab6b file from rosstat to tibble}
\usage{
tab6b_xls_convert(path_to_source, access_date = Sys.Date())
}
\arguments{
\item{path_to_source}{name of the original tab6b.xls file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts tab6b file from rosstat to tibble
}
\details{
Converts tab6b file from rosstat to tibble.
The structure is similar to tab5a.
}
\examples{
# no yet
}
