% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{tab9_xls_convert}
\alias{tab9_xls_convert}
\title{Converts tab9 file from rosstat to tibble}
\usage{
tab9_xls_convert(path_to_source, access_date)
}
\arguments{
\item{path_to_source}{name of the original tab9.xls file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts tab9 file from rosstat to tibble
}
\details{
Converts tab9 file from rosstat to tibble.
}
\examples{
# no yet
}
