% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_forecastors.R
\name{tbats_estimator}
\alias{tbats_estimator}
\title{tbats estimator}
\usage{
tbats_estimator(train_sample, predicted, options, predictors = "")
}
\arguments{
\item{train_sample}{tsibble with train sample}

\item{predicted}{name of the predicted variable}

\item{options}{character with options separated by comma
not yet supported}

\item{predictors}{not supported by tbats model}
}
\value{
fitted tbats model
}
\description{
Envelope function that estimates tbats model
}
\details{
Envelope function that estimates tbats model.
}
\examples{
train_sample = tsibble::tsibble(date = lubridate::ymd("2017-01-01") + months(0:99),
    y = arima.sim(n = 100, model = list(ar = 0.7)))
tbats_estimator(train_sample, "y", "")
}
