% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{tbats_fun}
\alias{tbats_fun}
\title{Do forecast using auto TBATS}
\usage{
tbats_fun(model_sample, h)
}
\arguments{
\item{model_sample}{preferably tsibble with "value" column}

\item{h}{forecasting horizon, is ignored}
}
\value{
auto TBATS model
}
\description{
Do forecast using auto TBATS
}
\details{
Do forecast using auto TBATS
}
\examples{
test = dplyr::tibble(date = as.Date("2017-01-01") + 0:9, value = rnorm(10))
tbats_fun(test, 1)
}
