% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_processing.R
\name{ts_2_tibble}
\alias{ts_2_tibble}
\title{Transform ts to tibble}
\usage{
ts_2_tibble(ts_data)
}
\arguments{
\item{ts_data}{ts object}
}
\value{
tsibble
}
\description{
Transform ts to tibble
}
\details{
Transform ts to tibble. The function takes univariate or multivariate
time series and transforms them to a tibble with a separate column indicating
a period of time.
Column with dates is called "date".
}
\examples{
test_ts = stats::ts(rnorm(100), start = c(2000, 1), freq = 12)
ts_2_tibble(test_ts)
}
