% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{uni_forecastor}
\alias{uni_forecastor}
\title{forecasts univariate model fit}
\usage{
uni_forecastor(fit, h = 1, test_sample = NULL)
}
\arguments{
\item{fit}{univariate model fit}

\item{h}{forecasting horizon}

\item{test_sample}{future regressors}
}
\value{
forecast object
}
\description{
forecasts univariate model fit
}
\details{
forecasts univariate model fit
}
\examples{
model = forecast::ets(rnorm(100))
uni_forecastor(model, h = 3)
}
