% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{uni_model_2_scalar_forecast}
\alias{uni_model_2_scalar_forecast}
\title{Extract one scalar forecast from univariate model}
\usage{
uni_model_2_scalar_forecast(model, h = 1, model_sample = NA)
}
\arguments{
\item{model}{univariate model}

\item{h}{forecasting horizon}

\item{model_sample}{ignored}
}
\value{
mean scalar forecast
}
\description{
Extract one scalar forecast from univariate model
}
\details{
Extract one scalar forecast from univariate model
}
\examples{
test = dplyr::tibble(date = as.Date("2017-01-01") + 0:9, value = rnorm(10))
tbats = tbats_fun(test, 1)
uni_model_2_scalar_forecast(tbats, h = 2)
}
