% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watchdog}
\alias{watchdog}
\title{Tibble of converters and parsers to download data}
\format{
A data frame with 7 columns
\describe{
  \item{url}{url to download and parse, may be empty}
  \item{file_raw}{name of the raw file after download, may be empty}
  \item{file_main}{name of the main file after parsing, usually csv}
  \item{processing}{name of the processing function}
  \item{univariate}{TRUE for univariate and FALSE for multivariate}
  \item{frequency}{observations per year}
  \item{comment}{comment in free form}
}
}
\usage{
watchdog
}
\description{
Tibble of converters and parsers to download data
}
\keyword{datasets}
