% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_operations.R
\name{write_csv2_cp1251}
\alias{write_csv2_cp1251}
\title{Write csv file for cyrillic MS office}
\usage{
write_csv2_cp1251(x, path)
}
\arguments{
\item{x}{data frame}

\item{path}{path to file}
}
\value{
results of write operations
}
\description{
Write csv file for cyrillic MS office
}
\details{
Write csv file for cyrillic MS office
Uses ";" as separator, CP1251 encoding, no column names.
}
\examples{
test = dplyr::tibble(x = c("Привет", "Пока"), z = c(1, 2))
write_csv2_cp1251(test, path = tempfile())
}
