% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_1_06_0_xlsx}
\alias{convert_1_06_0_xlsx}
\title{Converts Investment file from rosstat to tibble}
\usage{
convert_1_06_0_xlsx(
  path_to_source = "http://www.gks.ru/bgd/regl/b20_02/IssWWW.exe/Stg/d010/1-06-0.xlsx",
  access_date = Sys.Date()
)
}
\arguments{
\item{path_to_source}{name of the original 1-06-0.xls file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts Investment file from rosstat to tibble
}
\details{
Converts Investment file from rosstat to tibble
Written by Rifat Eniliev
}
\examples{
\donttest{
invest = convert_1_06_0_xlsx()
}
# convert_1_06_0_xlsx = function(path_to_source = "http://www.gks.ru/bgd/regl/b20_02/IssWWW.exe/Stg/d010/1-06-0.xlsx", access_date = Sys.Date()) {
#   data = rio::import(path_to_source)
#   data_vector = data[4:23, 3:6]  \%>\% t() \%>\% as.vector()
#   colnames(data_vector) = NULL
#   data_ts = stats::ts(data_vector, start = c(1999, 1), freq = 4)
#   data_tsibble = tsibble::as_tsibble(data_ts) \%>\% dplyr::rename(date = index, invest = value)
#   data_tsibble = dplyr::mutate(data_tsibble, access_date = access_date, date = as.Date(date))
#   check_conversion(data_tsibble)
#   return(data_tsibble)
# }
}
