% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_1_nn_doc}
\alias{convert_1_nn_doc}
\title{Converts 1-nn file from rosstat to tibble}
\usage{
convert_1_nn_doc(path_to_source, access_date = Sys.Date())
}
\arguments{
\item{path_to_source}{name of the original 1-nn.doc file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts 1-nn file from rosstat to tibble
}
\details{
Converts 1-nn file from rosstat to tibble.
Written by: Rifat Enileev
}
\examples{
# no yet
}
