% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_i_ipc_xlsx}
\alias{convert_i_ipc_xlsx}
\title{Converts I_ipc file from rosstat to tibble}
\usage{

  convert_i_ipc_xlsx(path_to_source = "http://www.gks.ru/free_doc/new_site/prices/potr/I_ipc.xlsx",
  access_date = Sys.Date())
}
\arguments{
\item{path_to_source}{name of the original I_ipc.xls file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts I_ipc file from rosstat to tibble
}
\details{
Converts I_ipc file from rosstat to tibble
}
\examples{
\donttest{
cpi = convert_i_ipc_xlsx()
}
}
