% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_urov_12kv_doc}
\alias{convert_urov_12kv_doc}
\title{Converts urov_12kv file from rosstat to tibble}
\usage{

  convert_urov_12kv_doc(path_to_source = "http://www.gks.ru/free_doc/new_site/population/urov/urov_12kv.doc",
  access_date = Sys.Date())
}
\arguments{
\item{path_to_source}{name of the original urov_12kv.doc file}

\item{access_date}{date of access is appended to every observation}
}
\value{
tibble
}
\description{
Converts urov_12kv file from rosstat to tibble
}
\details{
Converts urov_12kv file from rosstat to tibble.
Written by: Vladimir Omelyusik
}
\examples{
\donttest{
urok_12kv = convert_urok_12kv_doc()
}
}
