% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchdog.R
\name{download_statistics}
\alias{download_statistics}
\title{Download all statistics}
\usage{
download_statistics(path, watchdog, access_date = Sys.Date(),
  method = "curl", extra = "-L")
}
\arguments{
\item{path}{path to raw data folder}

\item{watchdog}{watchdog file
Structure of watchdog:
url - url of file to download, may be NA
file_raw - name for local raw file
file_main - name for local processed file
processing - name of processing function
univariate - TRUE/FALSE
frequency - 4/12/etc
comment - self explanatory :)}

\item{access_date}{access date}

\item{method}{for downloading files, passed to `download.file()`: "curl", "wget", "libcurl", "auto", "internal", "wininet"}

\item{extra}{options for downloading files, passed to `download.file()`: used for "wget" and "curl" methods}
}
\value{
downloads log
}
\description{
Download all statistics
}
\details{
Download all statistics
}
\examples{
path = tempdir()
mini_watchdog = tibble::tibble(url = "http://www.gks.ru/free_doc/new_site/prices/potr/I_ipc.xlsx",
    file_raw = "i_ipc.xlsx", file_main = "i_ipc.csv", processing = "convert_i_ipc_xlsx")
download_log_new = download_statistics(path, mini_watchdog)
}
