% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{forecast_2_scalar}
\alias{forecast_2_scalar}
\title{transforms forecast object into point forecast}
\usage{
forecast_2_scalar(fcst_object, h = 1)
}
\arguments{
\item{fcst_object}{complex forecast from model}

\item{h}{forecasting horizon}
}
\value{
point forecast, single number
}
\description{
transforms forecast object into point forecast
}
\details{
transforms forecast object into point forecast for common model types.
}
\examples{
fcst_object = forecast::forecast(rnorm(100), h = 6)
forecast_2_scalar(fcst_object, h = 3)
}
