% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_processing.R
\name{get_last_n_obs}
\alias{get_last_n_obs}
\title{Get last high frequency observations}
\usage{
get_last_n_obs(hf_data, lf_date, n_lags = 1, one_row = TRUE)
}
\arguments{
\item{hf_data}{high frequency data}

\item{lf_date}{low frequency date}

\item{n_lags}{number of lags to add}

\item{one_row}{output representation}
}
\value{
tsibble with high frequency lags added
}
\description{
Get last high frequency observations
}
\details{
Get last high frequency observations. The function cuts n_lags
observations from the table hf_data, that are from the same day or before 
as the date indicated by lf_date.  
If one_row is TRUE the result is stacked in a row, otherwise 
the result is a column.
}
\examples{
hf_data <- tibble::tibble(date = lubridate::ymd("2011-01-01") + 
           lubridate::days(0:60), observations = rnorm(61))
get_last_n_obs(hf_data, "2011-02-23", n_lags = 7, 
           one_row = FALSE)
}
