% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{make_report}
\alias{make_report}
\title{Produces report table}
\usage{
make_report(download_log, path)
}
\arguments{
\item{download_log}{metadata about downloaded series}

\item{path}{path to specific day}
}
\value{
data.frame
}
\description{
Produces table with additional information concerning downloaded data.
}
\details{
Several metrics to detect potentional download errors.
}
\examples{
\donttest{
path = "C:/CMF/Github repos/data/raw/"
path_day = paste0(path, "/", Sys.Date(), "/") # add current date to path
download_log_new = download_statistics(path, watchdog)
report_table <- make_report(download_log_new, path_day)
}
}
